Alright, I've been working on refactoring your Python imports and verifying the Docker configuration. Here's what I've done:
- First, I searched for `from backend.` imports, but didn't find any.
- Next, I corrected relative `..` imports in `agent/api.py`, `agentpress/plan_executor.py`, `utils/s3_upload_utils.py`, `utils/scripts/archive_inactive_sandboxes.py`, and `agentpress/response_processor.py` to use absolute paths from the `backend` root.
- Then, I made sure the content of `agentpress/task_types.py` was merged into `agentpress/api_models_tasks.py`. I also updated the imports in `tests/agentpress/test_plan_executor.py` and `tests/agentpress/test_task_state_manager.py` to use `api_models_tasks.py`, and then removed `agentpress/task_types.py`.
- Finally, I checked that the `backend` service configuration in your main `docker-compose.yaml` matches all the specified requirements, including `user`, `working_dir`, `command`, and `PYTHONPATH`.
