+++ backend/agentpress/plan_executor.py
@@ -16,14 +16,14 @@ import json
 import uuid
 import asyncio
 
-from ..services import redis
+from services import redis
 from .task_state_manager import TaskStateManager
 from .tool_orchestrator import ToolOrchestrator
 from .api_models_tasks import TaskState
 from .tool import ToolResult
 from .utils.json_helpers import format_for_yield
-from ..services.llm import make_llm_api_call
-from ..utils.logger import logger
+from services.llm import make_llm_api_call
+from utils.logger import logger
 
 class PlanExecutor:
     """
