+++ backend/agentpress/response_processor.py
@@ -16,13 +16,13 @@ import asyncio
 from datetime import datetime, timezone
 from typing import List, Dict, Any, Optional, AsyncGenerator, Tuple, Union, Callable, Literal
 from dataclasses import dataclass
-from ..utils.logger import logger
+from utils.logger import logger
 from .tool import ToolResult
 from .tool_orchestrator import ToolOrchestrator
 from .plan_executor import PlanExecutor
 from litellm import completion_cost
 from langfuse.client import StatefulTraceClient
-from ..services.langfuse import langfuse
+from services.langfuse import langfuse
 from .utils.json_helpers import (
     ensure_dict, ensure_list, safe_json_parse, 
     to_json_string, format_for_yield
