+++ backend/utils/s3_upload_utils.py
@@ -6,7 +6,7 @@ import base64
 import uuid
 from datetime import datetime
 from .logger import logger
-from ..services.supabase import DBConnection
+from services.supabase import DBConnection
 
 async def upload_base64_image(base64_data: str, bucket_name: str = "browser-screenshots") -> str:
     """Upload a base64 encoded image to Supabase storage and return the URL.
